requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.14.0
torch==2.1.1
transformers==4.35.2
python-multipart==0.0.6
pillow==10.1.0
plotly==5.18.0
streamlit==1.28.2
pytest==7.4.3
jupyter==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
nltk==3.8.1 